+++ src/js/notes/index.js
@@ -1,0 +1,12 @@
+import {
+  removeNote,
+  createNote,
+  updateNote,
+  toggleArchiveNote,
+} from "./notesCrud";
+export {
+  removeNote,
+  createNote,
+  updateNote,
+  toggleArchiveNote,
+};
