+++ src/js/render/index.js
@@ -1,0 +1,3 @@
+import renderItem from "./renderItem";
+
+export { renderItem };
