+++ src/js/state/index.js
@@ -1,0 +1,23 @@
+import {
+  notes,
+  addNoteItem,
+  removeNoteItem,
+  updateNoteItem,
+  getNoteItemById,
+  getArchivedNotesLength,
+  getNonArchivedNotesLength,
+  countStats,
+  renderStats,
+} from "./state";
+
+export {
+  notes,
+  addNoteItem,
+  removeNoteItem,
+  updateNoteItem,
+  getNoteItemById,
+  getArchivedNotesLength,
+  getNonArchivedNotesLength,
+  countStats,
+  renderStats,
+};
